requirements.txt
Flask==2.3.3
gunicorn==21.2.0
scikit-learn==1.4.2
pillow==10.3.0
pymongo==4.7.2
joblib==1.4.2
